backend/requirements.txt
Flask
Flask-Login
Flask-Cors
python-dotenv
paho-mqtt
flask-jwt-extended
werkzeug
flask_limiter
flask-sock